+++ CHANGELOG.md
@@ -11,6 +11,7 @@ be considered breaking changes.
 ### Fixed
 - A bug causing some funds to not be counted in balance calculations has been
   fixed.
+- JSON-RPC 1.x responses now use the expected HTTP error codes.
 
 ## [0.1.0-alpha.1] - 2025-09-18