+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 During the alpha period, no Semantic Versioning is followed; all releases should
 be considered breaking changes.
 
+## [0.1.0-alpha.2] - PLANNED
+
+### Fixed
+- A bug causing some funds to not be counted in balance calculations has been
+  fixed.
+- JSON-RPC 1.x responses now use the expected HTTP error codes.
+- JSON-RPC error codes now match zcashd more often.
+
 ## [0.1.0-alpha.1] - 2025-09-18
 
 Inital alpha release.